core-api: split route ref types into a separate modules
