+++ packages/core-api/src/routing/index.ts
@@ -22,6 +22,8 @@ export type {
   ExternalRouteRef,
 } from './types';
 export { FlatRoutes } from './FlatRoutes';
-export { createRouteRef, createExternalRouteRef } from './RouteRef';
+export { createRouteRef } from './RouteRef';
+export { createSubRouteRef } from './SubRouteRef';
+export { createExternalRouteRef } from './ExternalRouteRef';
 export type { RouteRefConfig } from './RouteRef';
 export { useRouteRef } from './hooks';
